requirements.txt
flask==3.1.2
python-dotenv==1.2.1
flasgger==0.9.7.1
boto3==1.42.50
flask-marshmallow==1.4.0
flask-compress==1.23
psycopg2-binary==2.9.11
redis==7.2.0
celery==5.6.2
gunicorn==25.1.0
flask-migrate==4.1.0
flask-sqlalchemy==3.1.1
flask-cors==6.0.2
flask-httpauth==4.8.0
pytz==2025.2